+++ filterEffects/filtersDB.js
@@ -4,7 +4,7 @@ import { FXLightningFilter } from "./filters/FXLightningFilter.js";
 import { FXPredatorFilter } from "./filters/FXPredatorFilter.js";
 import { FXOldFilmFilter } from "./filters/FXOldFilmFilter.js";
 import { FXBloomFilter } from "./filters/FXBloomFilter.js";
-import {FXFogFilter} from "./filters/FXFogFilter.js"
+// import {FXFogFilter} from "./filters/FXFogFilter.js"
 
 export const filtersDB = {
     lightning: FXLightningFilter,
@@ -13,5 +13,5 @@ export const filtersDB = {
     color: FXColorFilter,
     bloom: FXBloomFilter,
     oldfilm: FXOldFilmFilter,
-    fog: FXFogFilter
+    // fog: FXFogFilter
 }